+++ releaseNotes.md
@@ -1,6 +1,8 @@
 # Version 0.2
 
 - Increased max. number of results per query from 100 to 1000
+- Moved the functions to a separate functions module (functions.psm1)
+- Moved the CSV output from the ExecuteQuery function to a separate OutputCSV function
 
 # Version 0.1